CFG-174: Updating README: Running Unit Test Cases + Code Coverage
Running test cases with code coverage helps us to see what needs to be
removed. This is the local execution of the coverage implementation.

Currently, this is only working with the Unit test cases developed,
giving only 50% of coverage. The Integration tests need to use Docker to
simulate the execution against a Github repo.