CFG-675: Replacing yml with ymllint for validation with rules
The current python yml loader does NOT validate duplicate keys.

http://pyyaml.org/ticket/128

The yml linter is more powerful than the yml loader from python.
The list of rules that can be used is listed at
http://yamllint.readthedocs.io/en/latest/rules.html#.

The error message is as follows: [2:1: duplication of key "machine" in mapping (key-duplicates), 3:1: duplication of key "machine" in mapping (key-duplicates)]

*	modified:   validate_config_files.py
- Adding the linter config and linter from the test cases
https://github.com/edf-hpc/yamllint/blob/26b40707d9b19274ff445a5975b4ba89e81fd184/tests/common.py#L21-L22
- Loading the current rules that are simpler for the config server. Please
  refer to the rules list about what they are.
- Collecting the errors list and returning it
- Removing the try/catch for them as the lint returns the object